requirements.txt
google-generativeai
python-dotenv
tabulate
pytest
qdrant-client
sentence-transformers
pillow
pypdf
reportlab
streamlit